fix: when using container networking mode, ignore network overrides
When running a containerized step, the container is configured to use
the same network as the primary container for the task. Some
networking options, such as DNS configuration, are not supported when
using this networking mode, resulting in an error. Instead, filter
these options out, since the new container will implicity have them
configured via the networking on the primary container.